requirements.txt
pytest-playwright
playwright-stealth